fix: improve performance by tuning profiler disposal
Previously, if only running a single profile at a time, we would
recreate the CpuProfiler each time. This can cause a noticeable slowdown
with small profiles, up to 15x slower than the prior version.

With this change, we only recreate the profiler after a thousand
profiles have been captured. This avoids the constant overhead of
recreating the profiler while still occasionally recreating the profiler
to reclaim memory.